Move eventstream to /ecs-agent
